add docs for feluda core
