+++ feluda/factory/image_factory.py
@@ -14,8 +14,11 @@ log = logging.getLogger(__name__)
 
 
 class ImageFactory:
+    """Factory class for creating image objects from various sources."""
+
     @staticmethod
     def make_from_url(image_url: str) -> dict:
+        """Create an image object from a URL."""
         try:
             print("1", image_url)
             resp = requests.get(image_url, timeout=(3.05, 5))
@@ -33,6 +36,7 @@ class ImageFactory:
 
     @staticmethod
     def make_from_url_to_path(image_url: str) -> dict:
+        """Create an image object from a URL and save it to a path."""
         temp_dir = tempfile.gettempdir()
         temp_url = image_url.split("?", maxsplit=1)[0]
         file_name = temp_url.split("/")[-1]
@@ -48,6 +52,7 @@ class ImageFactory:
 
     @staticmethod
     def make_from_file_on_disk(image_path: str) -> dict:
+        """Create an image object from a file on disk."""
         with open(image_path, mode="rb") as file:
             image_bytes = file.read()
             image = Image.open(BytesIO(image_bytes))
@@ -60,10 +65,12 @@ class ImageFactory:
 
     @staticmethod
     def make_from_file_on_disk_to_path(image_path: str) -> dict:
+        """Create an image object from a file on disk and return its path."""
         return {"path": image_path}
 
     @staticmethod
     def make_from_file_in_memory(image_data: FileStorage) -> dict:
+        """Create an image object from a file in memory."""
         image_bytes = image_data.read()
         image = Image.open(BytesIO(image_bytes))
         image_array = np.array(image)
