+++ feluda/factory/s3_factory.py
@@ -4,6 +4,8 @@ import boto3
 
 
 class S3Factory:
+    """Factory class for handling S3 operations."""
+
     aws_access_key_id = os.getenv("AWS_ACCESS_KEY_ID")
     aws_secret_access_key = os.getenv("AWS_SECRET_ACCESS_KEY")
     aws_region = os.getenv("AWS_REGION")
@@ -19,6 +21,7 @@ class S3Factory:
     def download_file_from_s3(
         bucket_name: str, file_key: str, local_file_path: str
     ) -> None:
+        """Download a file from S3 to a local path."""
         try:
             S3Factory.s3.download_file(bucket_name, file_key, local_file_path)
             print(f"File {file_key} downloaded successfully!")
